pop up window fixed
